Move STB image manager inside GLFW module
