+++ glfw/src/Rendering/STBImageManager.hpp
