--- stb/CMakeLists.txt
@@ -1,3 +1,0 @@
-crimild_trace()
-
-add_subdirectory( src )
